CalculatedTypes Documentation

CalculatedTypes Documentation

Precompute expensive objects and/or generic interface for calculating parametrized objects.

Example

julia> @calculated_type(struct SumType{T}
            fieldA::T
            fieldB::T
        end, function calculate()
            fieldA + fieldB
        end,
        T
        );

julia> sum_type = SumType(1.0, 3.0);

julia> calc_sum_type = Calculated(sum_type);

julia> get_source(calc_sum_type)
SumType{Float64}(1.0, 3.0)

julia> get_value(calc_sum_type)
4.0

Functions

CalculatedTypes.Calculated — Method.

Calculated(obj::T)::CalculatedType{T}

Constructs the calculated version of obj::T, where the return type is a subtype of CalculatedType{T}. Must be implemented for every subtype of CalculatedType

source

CalculatedTypes.calculate — Method.

calculate(obj)

Return the value calculated for obj, to be stored in the CalculatedType's value field.

source

CalculatedTypes.get_source — Method.

get_source(calculated_object::CalculatedType)

Return the original object that generated calculated_object.

source

CalculatedTypes.get_value — Method.

get_value(calculated_object::CalculatedType)

Return the precomputed value stored by calculated_object.

source

CalculatedTypes.@calculated_type — Macro.

@calculated_type(type_def_expr, calculation_fn_expr=nothing, return_type=:Any)

Define a CalculatedType.

source